requirements.txt
asgiref==3.2.7
cffi==1.15.1
cryptography==41.0.3
Django==3.0.6
django-rest-knox==4.1.0
djangorestframework==3.11.0
pycparser==2.20
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
